requirements.txt
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==25.1.0
billiard==4.2.0
CacheControl==0.14.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
cssselect2==0.7.0
distlib==0.3.8
Django==5.0.1
django-cors-headers==4.4.0
django-cors-middleware==1.5.0
django-crispy-forms==2.3
django-dbbackup==4.2.1
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.9
et-xmlfile==1.1.0
filelock==3.15.4
firebase-admin==6.7.0
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
grpcio>=1.71.0
grpcio-status==1.71.0
gunicorn==22.0.0
html5lib==1.1
httplib2==0.22.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.0
lxml==5.2.2
MarkupSafe==2.1.5
msgpack==1.0.8
numpy==1.26.4
oscrypto==1.3.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pipenv==2024.0.1
platformdirs==4.2.2
prompt-toolkit==3.0.43
proto-plus>=1.23.0
protobuf>=5.26.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
PyMySQL==1.1.1
pyparsing==3.1.2
pypdf==4.3.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
razorpay==1.4.2
redis==5.0.3
referencing==0.36.2
reportlab==4.2.2
requests==2.31.0
rpds-py==0.23.1
rsa==4.9
setuptools==70.1.1
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.3.0
typing==3.7.4.3
typing_extensions==4.11.0
tzdata==2025.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.1
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.16
